refactor: 수동 매칭, 자동 매칭 요청 형식, 타입 통일
수동 매칭에는 description, departureTime이 더 있음

+++ src/main/java/com/gachtaxi/domain/matching/common/dto/request/AutoMatchingPostRequest.java
@@ -8,11 +8,13 @@ public record AutoMatchingPostRequest(
 //    String startName,
     String destinationPoint,
 //    String destinationName,
+
+    // 현재 사용하는 필드
     String startName,
     String destinationName,
     List<String> criteria,
-    List<Integer> members,
-    Integer expectedTotalCharge
+    List<Long> members,
+    int expectedTotalCharge
 ) {
 
   public List<Tags> getCriteria() {
